chore: Renamed 'plexapi' module to 'plex' to prevent import conflict
